fix datapack generation ignoring worlds when updating from <3.7.0
